Update BPG_Ball.uasset
I was pulling my hair out trying to figure out why the player was slowly detaching from the object when spinning, and how to fix it.

Answer: Its not the ball's fault
The players SnapToSpline was messing with the local position while it was attached to the ball, moving slightly slower when the plane was not aligned. More noticeable on corners.