Decoupling ABP from NightsLike
BP_NightsLike should not *require* the animation blueprint to function. The animation blueprint only serves to expand upon the state present in the pawn it is attached to.
Basically, anything on the player that used an ABP reference I instead moved to the ABP. As a consequence, the ABP maintains a hard reference to specifically BP_NightsLike to be able to retrieve the values it needs, but like, im like 90% sure thats fine.

Also the Launcher has a reference to NightsLike to make them spin. This implementation has been updated.

TODO: Find better solution for grabbables and be smarter about the launcher
Basically be smarter about things that affect the player that aren't the player